+++ src/test/css/deprecated-css-custom-property-syntax_test.ts
@@ -36,12 +36,12 @@ suite('deprecated-css-custom-property-syntax', () => {
 
   test('works in the trivial case', async() => {
     const warnings = await linter.lint([]);
-    assert.deepEqual(warnings, []);
+    assert.deepEqual([...warnings], []);
   });
 
   test('gives no warnings for a perfectly fine file', async() => {
     const warnings = await linter.lint(['perfectly-fine/polymer-element.html']);
-    assert.deepEqual(warnings, []);
+    assert.deepEqual([...warnings], []);
   });
 
   test('warns for the proper cases and with the right messages', async() => {
@@ -75,7 +75,7 @@ suite('deprecated-css-custom-property-syntax', () => {
     const warnings = await linter.lint(
         ['deprecated-css-custom-property-syntax/before-fixes.html']);
     const edits = warnings.filter((w) => w.fix).map((w) => w.fix!);
-    const loader = makeParseLoader(analyzer, await analyzer.analyze([]));
+    const loader = makeParseLoader(analyzer, warnings.analysis);
     const result = await applyEdits(edits, loader);
     assert.deepEqual(
         result.editedFiles.get(