+++ src/test/html/undefined-elements_test.ts
@@ -35,12 +35,12 @@ suite('undefined-elements', () => {
 
   test('works in the trivial case', async() => {
     const warnings = await linter.lint([]);
-    assert.deepEqual(warnings, []);
+    assert.deepEqual([...warnings], []);
   });
 
   test('gives no warnings for a perfectly fine file', async() => {
     const warnings = await linter.lint(['perfectly-fine/polymer-element.html']);
-    assert.deepEqual(warnings, []);
+    assert.deepEqual([...warnings], []);
   });
 
   test('finds undefined element references', async() => {