+++ src/test/polymer/content-to-slot-declarations_test.ts
@@ -34,13 +34,13 @@ suite(ruleId, () => {
 
   test('works in the trivial case', async() => {
     const warnings = await linter.lint([]);
-    assert.deepEqual(warnings, []);
+    assert.deepEqual([...warnings], []);
   });
 
   test('applies automatic-safe fixes', async() => {
     const warnings = await linter.lint([`${ruleId}/before-fixes.html`]);
     const edits = warnings.filter((w) => w.fix).map((w) => w.fix!);
-    const loader = makeParseLoader(analyzer);
+    const loader = makeParseLoader(analyzer, warnings.analysis);
     const result = await applyEdits(edits, loader);
     assert.deepEqual(
         result.editedFiles.get(`${ruleId}/before-fixes.html`),
@@ -60,7 +60,7 @@ suite(ruleId, () => {
         }
       }
     }
-    const loader = makeParseLoader(analyzer);
+    const loader = makeParseLoader(analyzer, warnings.analysis);
     const result = await applyEdits(edits, loader);
     assert.deepEqual(
         result.editedFiles.get(`${ruleId}/before-fixes.html`),